Split out project_tag.rs
